refactor: move factories to own namespace
