+++ src/Factory/TeaserFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Atoolo\GraphQL\Search\Resolver;
+namespace Atoolo\GraphQL\Search\Factory;
 
 use Atoolo\GraphQL\Search\Types\Teaser;
 use Atoolo\Resource\Resource;
