+++ src/Resolver/ResourceResolver.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Atoolo\GraphQL\Search\Resolver;
 
+use Atoolo\GraphQL\Search\Factory\DelegatingTeaserFactory;
 use Atoolo\GraphQL\Search\Types\Hierarchy;
 use Atoolo\GraphQL\Search\Types\Teaser;
 use Atoolo\Resource\Resource;
